requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pytz==2018.7
python-dotenv==0.8.2